Added
According to testing GET request to Logentries API returns a bunch of events but they are not sorted in a timeline manner. That is you might get newer events first and then the older ones. This makes searching for the starting point (which i used in the previous commit) for reading events useless. That is why this commit compares the events from previous request with current and sends to syslog only the new ones.
Also added timeout value between requests and amount of events pulled by one request as parameters for launching this script.